requirements.txt
autopep8==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.2.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
requests==2.28.2
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.14
watchdog==2.3.1
Werkzeug==2.2.3
